fix(autojump): fix strange behavior diff on unpacking

+++ home-manager/packages/shells/common.nix
@@ -9,7 +9,10 @@ let
     nativeBuildInputs = with pkgs; [ gzip ];
     installPhase = ''
       mkdir -p "$out/bin";
-      cp $src/autojump "$out/bin/autojump";
+      # strange error, on nixos the autojump is $src
+      # but on iprc the autoujump is located at $src/autojump
+      test -f $src/autojump && cp $src/autojump "$out/bin/autojump";
+      test -f $src && cp $src "$out/bin/autojump";
       chmod +x "$out/bin/autojump";
     '';
   };
